% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{calculateROC}
\alias{calculateROC}
\title{Calculate ROC attributes}
\usage{
calculateROC(x)
}
\arguments{
\item{x}{Train object generated by caret + scPred}
}
\value{
ROC object
}
\description{
Gets ROC object from a cell type
}
