% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{crossTab,scPred-method}
\alias{crossTab,scPred-method}
\title{Gets contingency table}
\usage{
\S4method{crossTab}{scPred}(object, true, pred = NULL, fill = 0,
  prop = TRUE, digits = 2)
}
\arguments{
\item{object}{\code{scPred} object}

\item{true}{Column name in \code{predMeta} slot that corresponds to the true known classes}

\item{pred}{Column name in \code{predMeta} slot that corresponds to the predicted classes
if they  have been assigned independently from the \code{scPredict()} function}

\item{fill}{Value to fill contingency table ff unique cell classes from the true and the 
predicted columns do not match.}

\item{prop}{Return proportions or counts? Default: proportions}

\item{digits}{If proportions are returned, number of digits to round numbers}
}
\value{
A contingency table
}
\description{
If the \code{predMeta} slot contains a column with the true classes of the cells,
builds a contingency table by using this column as reference and comparing it to the predicted classes
obtained with \code{scPredict}. If an independent column with predicted classes is in the prediction
metadata, this column instead of the default classes assigned by \code{scPred} can be provided using the
\code{pred} parameter.
}
\author{
José Alquicira Hernández
}
