% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{crossTab,Seurat-method}
\alias{crossTab,Seurat-method}
\title{Gets contingency table}
\usage{
\S4method{crossTab}{Seurat}(
  object,
  true = NULL,
  pred = NULL,
  output = c("counts", "fraction", "prop"),
  digits = 2
)
}
\arguments{
\item{object}{\code{Seurat} object}

\item{true}{Column name in \code{meta.data} slot that corresponds to the true known classes}

\item{pred}{Column name in \code{meta.data} slot that corresponds to the predicted classes
if they  have been assigned independently from the \code{scPredict()} function}

\item{output}{Return counts, fraction, or proportions? Default: counts}

\item{digits}{If proportions are returned, number of digits to round numbers}
}
\value{
A contingency table
}
\description{
Creates a cross table using two columns from the metadata
}
\author{
Jose Alquicira Hernandez
}
