% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSubspaceSeurat.R
\name{.alignSubspaceSeurat}
\alias{.alignSubspaceSeurat}
\title{Align low-dimensional space using Seurat algorithm}
\usage{
.alignSubspaceSeurat(object, reduction.type = "pca.scpred",
  grouping.var = "dataset_dummy", dims.align,
  num.possible.genes = 2000, num.genes = 30, show.plots = FALSE,
  verbose = TRUE, ...)
}
\description{
Uses the manifold-alignment Seurat algorithm to align the training eigenspace and the prediction projection
See ?AlignSubspace() for more details. Note: this helper function is a modified version from Seurat.
}
\author{
José Alquicira Hernández
}
