% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{.plot_probabilities}
\alias{.plot_probabilities}
\title{Plot training probabilities}
\usage{
.plot_probabilities(object, size = 0.8)
}
\arguments{
\item{object}{Seurat or scPred object}

\item{size}{Point size for each cell}
}
\value{
Plot with the probability distribution for each cell type
}
\description{
Plots all training probabilities for each cell type
}
