% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenDecompose.R
\name{eigenDecompose}
\alias{eigenDecompose}
\title{Eigendecompose gene expression matrix via SVD}
\usage{
eigenDecompose(expData, pseudo = TRUE, args = NULL, decimals = 2)
}
\arguments{
\item{pseudo}{Whether to perform a \code{log2(data + 1)} transformation on expression data}

\item{args}{List with extra arguments provided to \code{prcomp()} function.}

\item{decimals}{Number of decimals for explained variance results}

\item{expdata}{A matrix object with cells as rows and genes (loci) as columns. Unique row names must be provided}
}
\value{
A eigenPred object with three filled slots
\itemize{
\item \code{prcomp}: results from \code{prcomp} function
\item \code{expVar}: explained variance by each principal component
\item \code{pseudo}: \code{TRUE} if a pseudo-log2 transformation was performed
}
}
\description{
Performs principal component analysis from a gene expression matrix. Data is centered and scaled by default.
}
\examples{

# Eigendecompose gene expression matrix
res <- eigendecompose(expTrain)
}
\author{
José Alquicira Hernández
}
\keyword{decomposition,}
\keyword{pca}
\keyword{singular}
\keyword{value}
