% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenPredict.R
\name{eigenPredict}
\alias{eigenPredict}
\title{Predict classes of a new dataset using a trained model}
\usage{
eigenPredict(object, newData, threshold = 0.7)
}
\arguments{
\item{trainData}{An \code{scPred} object with metadata and informative features obtained.}

\item{referenceData}{A matrix object with cells as rows and genes (loci) as columns obtained with \code{projectNewData} function}

\item{trainedModel}{A \code{train} object returned by \code{trainModel} function}
}
\description{
Predicts cell classes for a new dataset based on a training model, a reference \code{eigenPred} object
}
\author{
José Alquicira Hernández
}
\keyword{new,}
\keyword{prediction,}
\keyword{test,}
\keyword{validation}
