% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_classifiers,scPred-method}
\alias{get_classifiers,scPred-method}
\title{Get classification models}
\usage{
\S4method{get_classifiers}{scPred}(object)
}
\arguments{
\item{object}{\code{scPred} object}
}
\value{
A list of \code{train} objects
}
\description{
Accessor function to retrieve trained models from scPred objects
}
