% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_classifiers,Seurat-method}
\alias{get_classifiers,Seurat-method}
\title{Get classification models}
\usage{
\S4method{get_classifiers}{Seurat}(object)
}
\arguments{
\item{object}{\code{Seurat} object}
}
\value{
A list of \code{train} objects
}
\description{
Accessor function to retrieve scPred models from Seurat objects
}
