% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_classifiers}
\alias{get_classifiers}
\title{Get classification models}
\usage{
get_classifiers(object)
}
\arguments{
\item{object}{\code{Seurat} or scPred object}
}
\value{
A list of \code{train} objects
}
\description{
Accessor function to retrieve classification models from Seurat 
and scPred objects
}
