% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_metadata,scPred-method}
\alias{get_metadata,scPred-method}
\title{Get metadata from scPred object}
\usage{
\S4method{get_metadata}{scPred}(object)
}
\value{
A dataframe including the cell barcodes and prediction variable 
(cell type labels)
}
\description{
Accessor function to retrieve metadata from scPred object
}
