% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from scPred object}
\usage{
get_metadata(object)
}
\value{
A dataframe including the cell barcodes and prediction variable 
(cell type labels)
}
\description{
Accessor function to retrieve metadata from scPred object
}
