% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_probabilities,scPred-method}
\alias{get_probabilities,scPred-method}
\title{Get training probabilities}
\usage{
\S4method{get_probabilities}{scPred}(object)
}
\arguments{
\item{object}{scPred object}
}
\value{
A data frame with all cell-type probabilities associated to each cell
}
\description{
Gets training probabilities for each cell type
}
