% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_probabilities}
\alias{get_probabilities}
\title{Get training probabilities}
\usage{
get_probabilities(object)
}
\arguments{
\item{object}{Seurat or scPred object}
}
\value{
A data frame with all cell-type probabilities associated to each cell
}
\description{
Gets training probabilities for each cell type
}
