% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_scpred,Seurat-method}
\alias{get_scpred,Seurat-method}
\title{Get scPred object}
\usage{
\S4method{get_scpred}{Seurat}(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
scPred object
}
\description{
Accessor function to retrieve scPred models from Seurat objects
}
