% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_scpred}
\alias{get_scpred}
\title{Get scPred object}
\usage{
get_scpred(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
scPred object
}
\description{
Accessor function to retrieve scPred models from Seurat objects
}
