% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureSpace.R
\name{getFeatureSpace}
\alias{getFeatureSpace}
\title{Get discriminant feature space}
\usage{
getFeatureSpace(object, pVar, varLim = 0.01, correction = "fdr", sig = 0.05)
}
\arguments{
\item{object}{A \code{seurat} object}

\item{pVar}{Column in \code{meta.data} slot containing the cell-type labels of each single cell}

\item{varLim}{Threshold used to filter principal components based on their individual variance explained.}

\item{correction}{Multiple testing correction method used. Default: false discovery rate. See \code{p.adjust} function}

\item{sig}{Significance level to determine principal components explaining class identity}
}
\value{
An \code{Seurat} object along with a \code{scPred} object stored in the \code{@misc} slot 
containing a data.frame of significant features with the following columns:
\itemize{
\item PC: Principal component
\item pValue: p-value obtained from Wilcoxon rank sum test
\item pValueAdj: Adjusted p-value according to \code{correction} parameter
\item expVar: Explained variance for each principal component
\item cumExpVar: All principal components are ranked according to their significance and variance explained. 
This column contains the cumulative variance explained across the ranked principal components
}
}
\description{
Given a prediction variable, finds a feature set of class-informative principal components that
explain variance differences between cell types.
}
\examples{
pbmc_small <- getFeatureSpace(pbmc_small, "RNA_snn_res.0.8")

}
\author{
Jose Alquicira Hernandez
}
\keyword{features}
\keyword{informative,}
\keyword{significant,}
