% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInformativePCs.R
\name{getInformativePCs}
\alias{getInformativePCs}
\title{Get informative principal components}
\usage{
getInformativePCs(object, pVar, rep = 100, seed = NULL, varLim = 0.01,
  correction = "fdr", sig = 0.05)
}
\arguments{
\item{object}{An \code{eigenPred} object}

\item{pVar}{Prediction variable corresponding to a column in \code{metadata} slot}

\item{rep}{Number of random samples to assess stable principal components}

\item{seed}{Seed to create random samples}

\item{varLim}{Threshold to filter principal components based on variance explained.}

\item{correction}{Multiple testing correction method. Default: false discovery rate. See \code{p.adjust} function}

\item{sig}{Significance level to determine principal components explaining class identity}
}
\value{
A \code{eigenPred} object with three additional filled slots:
\itemize{
\item \code{features}: A data frame with significant principal components the following information:
\itemize{
\item PC: Principal component
\item Freq: Frequency of occurence of the principal component over a number of random samples from the PCA matrix
\item expVar: Explained variance by the principal component
\item cumExpVar: All principal components are ranked accoriding to their frequency of ocurrence and their variance explained. 
This column contains the cumulative variance explained across the ranked principal components
}
\item \code{pVar}: Column name from metadata to use as the variable to predict using
the informative principal components. Informative principal components are selected based on this variable.
\item \code{rep}: Number of random samples to determine stable principal components
}
}
\description{
Given a prediction variable, finds a set of class-informative principal components that are significant across n subsets of cells from the 
PCA matrix. A Wilcoxon rank sum test is used to determine a difference between the value distributions of classes from the prediction variable.
}
\author{
José Alquicira Hernández
}
\keyword{features}
\keyword{informative,}
\keyword{significant,}
