% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getLoadings,eigenPred-method}
\alias{getLoadings,eigenPred-method}
\title{Get loadings matrix}
\usage{
\S4method{getLoadings}{eigenPred}(object)
}
\description{
Gets rotation matrix (right singular vectors) from a \code{eigenPred} object
}
