% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getLoadings,scPred-method}
\alias{getLoadings,scPred-method}
\title{Get loadings matrix}
\usage{
\S4method{getLoadings}{scPred}(object)
}
\description{
Gets rotation matrix (right singular vectors) from a \code{scPred} object
}
