% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getPCA,eigenPred-method}
\alias{getPCA,eigenPred-method}
\title{Get principal components}
\usage{
\S4method{getPCA}{eigenPred}(object)
}
\description{
Gets matrix with principal components from a \code{eigenPred} object
}
