% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{getPCA,scPred-method}
\alias{getPCA,scPred-method}
\title{Get principal components}
\usage{
\S4method{getPCA}{scPred}(object)
}
\description{
Gets matrix with principal components from a \code{scPred} object
}
