% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getPredictions,scPred-method}
\alias{getPredictions,scPred-method}
\title{Get training probabilities}
\usage{
\S4method{getPredictions}{scPred}(object)
}
\description{
Gets prediction probabilities for each cell class
}
