% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getProbabilities,Seurat-method}
\alias{getProbabilities,Seurat-method}
\title{Get training probabilities}
\usage{
\S4method{getProbabilities}{Seurat}(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A data frame with all cell-type probabilities associated to each cell
}
\description{
Gets training probabilities for each cell type
}
