% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getProbabilities}
\alias{getProbabilities}
\title{Get training probabilities}
\usage{
getProbabilities(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
A data frame with all cell-type probabilities associated to each cell
}
\description{
Gets training probabilities for each cell type
}
