% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getROC,Seurat-method}
\alias{getROC,Seurat-method}
\title{Get ROC objects for all cell types}
\usage{
\S4method{getROC}{Seurat}(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Prints scPred object
}
\description{
Gets ROC objects for all cell types from a Seurat/scPred object
}
