% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getROC}
\alias{getROC}
\title{Get ROC objects for all cell types}
\usage{
getROC(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
Prints scPred object
}
\description{
Gets ROC objects for all cell types from a Seurat/scPred object
}
