% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getThresholds,Seurat-method}
\alias{getThresholds,Seurat-method}
\title{Get probability thresholds}
\usage{
\S4method{getThresholds}{Seurat}(object, tolerance = 0.05, nmads = 3)
}
\arguments{
\item{object}{Seurat object}

\item{tolerance}{Probability tolerance value for threshold. This value is substracted or added from
the computed probability thresholds for the cell type of interest and the remaining cell types respectively.
Useful when applying using a new dataset where computed probabilities may be slightly lower}

\item{nmads}{Number of median absolute deviations to use}
}
\value{
A data frame of 2 x n cell types. Positive thresholds refer to the cell type of interest and
negative to other cell types in a binary classification setting
}
\description{
Calculates probability thresholds based on median absolute deviations
}
