% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTrainResults,scPred-method}
\alias{getTrainResults,scPred-method}
\title{Get training probabilities}
\usage{
\S4method{getTrainResults}{scPred}(object)
}
\description{
Gets training probabilities for each trained model
}
