% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{metadata<-}
\alias{metadata<-}
\title{Set metadata}
\usage{
metadata(object) <- value
}
\description{
Sets metadata to a \code{scPred} object. Metadata must be a dataframe
\itemize{
\item row names: ids matching the column names of the expression matrix
\item columns: associated metadata such as cell type, conditions, sample or, batch. 
}
}
