% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_2}
\alias{pbmc_2}
\title{A random sample of PBMCs from individual 2}
\format{
A Seurat object with count data
}
\source{
\url{https://academic.oup.com/nargab/article/2/2/lqaa034/5836690}
}
\usage{
pbmc_2
}
\description{
A subsetted version of sample 2 from: Comparative performance of the BGI and 
Illumina sequencing technology for single-cell RNA-sequencing
}
\keyword{datasets}
