% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_probabilities,scPred-method}
\alias{plot_probabilities,scPred-method}
\title{Plot training probabilities}
\usage{
\S4method{plot_probabilities}{scPred}(object, size = 0.8)
}
\arguments{
\item{object}{scPred object}

\item{size}{Point size for each cell}
}
\value{
Plot with the probability distribution for each cell type
}
\description{
Plots all training probabilities for each cell type
}
