% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plotEigen,eigenPred-method}
\alias{plotEigen,eigenPred-method}
\title{Plot PCA}
\usage{
\S4method{plotEigen}{eigenPred}(object, group = NULL, pc = c(1, 2),
  geom = c("both", "points", "density_2d"))
}
\description{
Plots PCA
}
