% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotEigen}
\alias{plotEigen}
\title{Plot PCA}
\usage{
plotEigen(object, group = NULL, pc = c(1, 2), geom = c("points",
  "density_2d", "both"))
}
\description{
Plots PCA with raw results or group by a variable
}
