% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotExp,scPred-method}
\alias{plotExp,scPred-method}
\title{Plot gene expression data}
\usage{
\S4method{plotExp}{scPred}(object, gene, pc = c(1, 2), low = "gray", high = "red")
}
\arguments{
\item{object}{\code{scPred} object}

\item{gene}{Gene id. Must match one of the row names in the gene expression matrix used as input}

\item{pc}{Vector of length 2 with the two dimensions to be plotted}

\item{low}{Color for low gene expression}

\item{high}{Color for high gene expression}
}
\description{
Plots a PCA and projection with the gene expression values for a given gene
}
