% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{plotFeatures,eigenPred-method}
\alias{plotFeatures,eigenPred-method}
\title{Generate diagnostic plot}
\usage{
\S4method{plotFeatures}{eigenPred}(object)
}
\description{
Plots a graph showing the frequency of each principal component across a number of random samples and the cumulative variance
explained by them. This plot may be used to select stable principal components to be used to train a prediction model.
}
