% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotFeatures}
\alias{plotFeatures}
\title{Generate diagnostic plot}
\usage{
plotFeatures(object)
}
\description{
Plots a graph showing the frequency of each principal component across a number of random samples and the cumulative variance
explained by them. This plot may be used to select stable principal components to be used to train a prediction model.
}
