% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotLoadings}
\alias{plotLoadings}
\title{Plot loadings}
\usage{
plotLoadings(object, pc = 1, n = 10)
}
\arguments{
\item{object}{A \code{scPred} object}

\item{pc}{The number of the principal component to be plotted}

\item{n}{Top `n` variable genes to plot. Notice that the number of genes plotted is n*2 as both
negative and positive loadings are considered}
}
\description{
Plot loading values for any given principal component in a \code{scPred} object
}
