% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotPredProbs,scPred-method}
\alias{plotPredProbs,scPred-method}
\title{Plot prediction probabilities}
\usage{
\S4method{plotPredProbs}{scPred}(object, facet = NULL)
}
\arguments{
\item{object}{\code{scPred} object}

\item{facet}{Column name in \code{predMeta} slot
if they have been assigned independently from the \code{scPredict()} function}
}
\value{
A distribution probability plot divided into facets representing the groups contained in the
provided facet column name and colors to the corresponding probabilities for each class. If no grouping
variable is provided to the \code{facet} parameter, distributions are divided according to the predicted classes
by default
}
\description{
Plots the probability distributions according to a grouping variable.
}
\author{
José Alquicira Hernández
}
