% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotProbs,Seurat-method}
\alias{plotProbs,Seurat-method}
\title{Plot training probabilities}
\usage{
\S4method{plotProbs}{Seurat}(object, cost = 1, class_weight = TRUE, rug = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{cost}{The relative cost of a false negative classification (as compared with a false positive classification)}

\item{class_weight}{Use the proportion of each cell type vs others as the prevalance
(n.cases/(n.controls+n.cases)).}

\item{rug}{Add rug geom to plot}
}
\value{
Plot with the probability distribution for each cell type
}
\description{
Plots all training probabilities for each cell type
}
