% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plotROC}
\alias{plotROC}
\title{Plot ROC curve}
\usage{
plotROC(object, cost = 1, class_weight = TRUE, rug = FALSE)
}
\arguments{
\item{object}{Seurat object}

\item{cost}{The relative cost of a false negative classification (as compared with a false positive classification)}

\item{class_weight}{Use the proportion of each cell type vs others as the prevalance
(n.cases/(n.controls+n.cases)).}
}
\value{
Plot with the ROC curve for each cell type
}
\description{
Plots a ROC curve for each cell type using the training probabilities
}
