% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_utils.R
\name{plotTrainProbs}
\alias{plotTrainProbs}
\title{Plot training probabilities for all models}
\usage{
plotTrainProbs(object, ...)
}
\arguments{
\item{object}{An \code{scPred} or \code{seurat} object}

\item{...}{Extra arguments for \code{plot_grid} function}
}
\value{
A plot with the probability distributions. Each panel corresponds to a trained model (cell class)
and the positive class to the class of interest
}
\description{
Plots the probability distributions for each model.
}
\author{
José Alquicira Hernández
}
