% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.scPred}
\alias{print.scPred}
\title{show}
\usage{
\method{print}{scPred}(string)
}
\arguments{
\item{string}{list with output strings}
}
\value{
Prints scPred object
}
\description{
Displays scPred object info
}
