% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_query.R
\name{project_query}
\alias{project_query}
\title{Project query data onto PCA reference}
\usage{
project_query(
  new,
  reference,
  max.iter.harmony = 20,
  recompute_alignment = TRUE,
  seed = 66,
  ...
)
}
\arguments{
\item{new}{A Seurat object containing cells to be projected}

\item{reference}{A \code{Seurat} object after running \code{getFeatureSpace()}}

\item{max.iter.harmony}{Maximum number of rounds to run Harmony. One round of Harmony involves one clustering and one correction step.}

\item{recompute_alignment}{Recompute alignment? Useful if \code{scPredict()} has already been run}

\item{seed}{Numeric seed for harmony}

\item{...}{Further arguments to \code{HarmonyMatrix()}}
}
\value{
A Seurat object with two new reductions: 
\itemize{
\item \code{scpred}: Aligned data using harmony
\item \code{scpred_projection}: Raw projection using reference loadings
}
}
\description{
Projects data from Seurat object onto a pre-computed PCA
}
\author{
José Alquicira Hernández
}
\keyword{new,}
\keyword{prediction,}
\keyword{test,}
\keyword{validation}
