% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPredSeurat.R
\name{scPredSeurat}
\alias{scPredSeurat}
\title{Perform predictions via Seurat integration}
\usage{
scPredSeurat(reference, new, threshold = NULL, weight = TRUE, ...)
}
\arguments{
\item{reference}{A \code{Seurat} object used as reference to perform predictions}

\item{new}{A \code{Seurat} object from which cells will be classified}

\item{threshold}{Threshold used for probabilities to classify cells into classes}

\item{weight}{Whether output probabilities are averaged across models for every cell.}

\item{...}{Extra arguments passed to \code{FindTransferAnchors()}}
}
\value{
A matrix:.
\itemize{
\item Probabilities of each cell type for all cells in the \code{new} object
\item \code{prediction} Cell type predicted based on probability threshold
\item \code{generic} Cell type predicted based on maximum probability acroos all models
}
}
\description{
Integrates data using Seurat and performs predictions based on informative features
identified by \code{scPred}
}
\examples{

res <- scPredSeurat(reference = integrated, new = query)

}
\author{
José Alquicira Hernández
}
