% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{show,eigenPred-method}
\alias{show,eigenPred-method}
\title{show}
\usage{
\S4method{show}{eigenPred}(object)
}
\description{
Generic display function for \linkS4class{eigenPred} objects. Displays summary of the
object such as number of cells, genes, significant features.
}
