% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,scPred-method}
\alias{show,scPred-method}
\title{show}
\usage{
\S4method{show}{scPred}(object)
}
\description{
Generic display function for \linkS4class{scPred} objects. Displays summary of the
object such as number of cells, genes, significant features.
}
