% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignDataset.R
\name{alignDataset}
\alias{alignDataset}
\title{Align train and test datasets}
\usage{
alignDataset(object, prediction)
}
\description{
Aligns train and test datasets in a low dimensional space using the dynamic time warping alogorithm implemented in Seurat
}
