% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenDecompose.R
\name{eigenDecompose}
\alias{eigenDecompose}
\title{Eigendecompose gene expression matrix}
\usage{
eigenDecompose(expData, n = 10, pseudo = TRUE)
}
\arguments{
\item{n}{Number of principal components to be computed}

\item{pseudo}{Whether to perform a \code{log2(data + 1)} transformation on expression data}

\item{expdata}{A matrix object with cells as rows and genes (loci) as columns. Unique row names for genes must be provided}
}
\value{
A scPred object with three filled slots
\itemize{
\item \code{prcomp}: results from \code{prcomp} function
\item \code{expVar}: explained variance by each principal component
\item \code{pseudo}: \code{TRUE} if a pseudo-log2 transformation was performed
}
}
\description{
Performs principal component analysis from a gene expression matrix. Data is centered and scaled by default.
}
\examples{

# Eigendecompose gene expression matrix
object <- eigendecompose(expTrain, n = 25)
}
\author{
José Alquicira Hernández
}
\keyword{decomposition,}
\keyword{pca}
\keyword{singular}
\keyword{value}
