% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectNewData.R
\name{projectNewData}
\alias{projectNewData}
\title{Project new data into training principal components}
\usage{
projectNewData(newData, referenceData)
}
\arguments{
\item{newData}{A matrix object with cells as rows and genes (loci) as columns}

\item{referenceData}{An \code{scPred} object}
}
\value{
A data frame with the projection
}
\description{
Projects a new dataset into the principal components obtained from a training dataset
}
\author{
José Alquicira Hernández
}
\keyword{projection}
\keyword{test,}
\keyword{validation,}
