% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{scPred}
\alias{scPred}
\alias{scPred-class}
\title{Definition of 'scPred' class}
\description{
An S4 class to containing features, dimensionality reduction information, and trained models
}
\section{Slots}{

\describe{
\item{\code{pvar}}{Column name from metadata to use as the variable to predict using}

\item{\code{metadata}}{A data frame to store metadata including prediction variable (pvar)}

\item{\code{features}}{A data frame with the following information:
\itemize{
\item feature: Eigenvector (e.g. principal component)
\item pValue: Significance value from a wilcoxon test 
\item pValueAdj: Adjusted p-value for multiple testing
}}

\item{\code{loadings}}{Gene loadings}

\item{\code{scaling}}{Means and standard deviation to center and standardize data}

\item{\code{reduction}}{Dimensionality reduction name}

\item{\code{reduction_key}}{Dimensionality reduction name key}

\item{\code{train}}{A list with all trained models using the \code{caret} package. Each model correspond to a cell type}

\item{\code{mist}}{A list to store extra information and for developing testing}
}}

