% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPredict.R
\name{scPredict}
\alias{scPredict}
\title{Predict cell classes from a new dataset using a trained model}
\usage{
scPredict(reference, new, threshold = NULL, weight = TRUE, ...)
}
\arguments{
\item{reference}{An \code{Seurat} object with trained model(s) using \code{scPred}.}

\item{new}{A seurat object containing cells to be classified}

\item{threshold}{Threshold used for probabilities to classify cells into classes}

\item{weight}{If TRUE, probabilities from binary classifiers are weighted to sum 1.}
}
\value{
A Seurat object with addtional metadata columns with prediction probabilities associated to each class, a \code{prediction} column, 
indicating the classification based on the provided threshold and a \code{generic_class} column without "unassigned" labels.
}
\description{
Predicts cell classes for a new dataset based on trained model(s)
}
\examples{


new <- scPredict(reference, new)

}
\author{
José Alquicira Hernández
}
\keyword{new,}
\keyword{prediction,}
\keyword{test,}
\keyword{validation}
