% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPredict.R
\name{scPredict}
\alias{scPredict}
\title{Predict cell classes from a new dataset using a trained model}
\usage{
scPredict(object, newData, threshold = 0.9)
}
\arguments{
\item{object}{An \code{scPred} object with metadata and informative features and trained model(s).}

\item{newData}{A matrix object with cells as rows and genes (loci) as columns}

\item{threshold}{Threshold used for probabilities to classify cells into classes}
}
\value{
A data frame with prediction probabilities associated to each class and a \code{predClass} column, 
indicating the classification based on the provided threshold
}
\description{
Predicts cell classes for a new dataset based on a training model, a reference \code{eigenPred} object
}
\examples{

# Get class probabilities for cells in a new dataset. A 'predClass' columns with 
cell classes is returned depending on the threshold parameter.

prediction <- scPredict(object = object, newData = expTest, threshold = 0.9)

}
\author{
José Alquicira Hernández
}
\keyword{new,}
\keyword{prediction,}
\keyword{test,}
\keyword{validation}
