% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scPredict.R
\name{scPredict}
\alias{scPredict}
\title{Predict cell classes from a new dataset using a trained model}
\usage{
scPredict(
  new,
  reference,
  threshold = 0.75,
  tolerance = 0.1,
  max.iter.harmony = 20,
  recompute_alignment = TRUE,
  reference_scaling = TRUE,
  ...
)
}
\arguments{
\item{new}{A seurat object containing cells to be classified}

\item{reference}{An \code{Seurat} object with trained model(s) using \code{scPred} or an \code{scPred} object}

\item{threshold}{Threshold used for probabilities to classify cells into classes}

\item{tolerance}{Probability tolerance value for threshold. This value is substracted or added from
the computed probability thresholds for the cell type of interest and the remaining cell types respectively.}

\item{max.iter.harmony}{Maximum number of rounds to run Harmony. One round of Harmony involves one clustering and one correction step.}

\item{recompute_alignment}{Recompute alignment? Useful if \code{scPredict()} has already been run}

\item{reference_scaling}{Scale new dataset based on means and stdevs from reference dataset before alignment. Otherwise, data will be independently scaled.}
}
\value{
A Seurat object with addtional metadata columns with prediction probabilities associated to each class, a \code{prediction} column, 
indicating the classification based on the provided threshold and a \code{generic_class} column without "unassigned" labels.
}
\description{
Predicts cell classes for a new dataset based on trained model(s)
}
\author{
José Alquicira Hernández
}
\keyword{new,}
\keyword{prediction,}
\keyword{test,}
\keyword{validation}
