% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainModel.R
\name{trainModel}
\alias{trainModel}
\title{Train a prediction model}
\usage{
trainModel(
  object,
  model = "svmRadial",
  preProcess = c("center", "scale"),
  resampleMethod = "cv",
  number = 5,
  seed = 66,
  tuneLength = 3,
  metric = c("ROC", "PR", "Accuracy", "Kappa"),
  returnData = FALSE,
  savePredictions = "final",
  allowParallel = FALSE,
  reclassify = NULL
)
}
\arguments{
\item{object}{An \code{Seurat} or \code{scPred} object after running 
\code{getFeatureSpace}}

\item{model}{Classification model supported via \code{caret} package. A list of all models can be found here:}

\item{preProcess}{A string vector that defines a pre-processing of the predictor data. Current possibilities are 
"BoxCox", "YeoJohnson", "expoTrans", "center", "scale", "range", "knnImpute", "bagImpute", "medianImpute", 
"pca", "ica" and "spatialSign". The default is "center" and "scale. See preProcess and trainControl on the 
procedures and how to adjust them
https://topepo.github.io/caret/available-models.html
Default: support vector machine with radial kernel}

\item{resampleMethod}{Resample model used in \code{trainControl} function from \code{caret}. 
Default: K-fold cross validation}

\item{number}{Number of iterations for resample method. See \code{trainControl} function}

\item{seed}{Numeric seed for resample method. Fixed to ensure reproducibility}

\item{tuneLength}{An integer denoting the amount of granularity in the tuning parameter grid. 
By default, this argument is the number of levels for each tuning parameters that should be generated by train.
See `?caret::train` documentation}

\item{metric}{Performance metric to be used to select best model: `ROC` (area under the ROC curve), 
`PR` (area under the precision-recall curve), `Accuracy`, and `Kappa`}

\item{returnData}{If \code{TRUE}, training data is returned within \code{scPred} object.}

\item{savePredictions}{Specifies the set of hold-out predictions for each resample that should be
returned. Values can be either "all", "final", or "none".}

\item{allowParallel}{Allow parallel processing for resampling?}

\item{reclassify}{Cell types to reclassify using a different model}
}
\value{
A list of \code{train} objects for each cell class (e.g. cell type). See \code{train} function for details.
}
\description{
Trains a prediction model from an \code{scPred} object stored in a \code{Seurat} object
}
\author{
Jose Alquicira Hernandez
}
\keyword{model}
\keyword{train,}
